Infraestructura/ExporterWS/requirements.txt
tornado==4.3
python-dateutil==2.5.3
uuid==1.30
pyjwt==2.1.0
jinja2==3.0.1
pdfkit==0.6.1
qrcode==7.3.1
pillow==8.3.1